Extract order utilities
